♻️ (templates): Extract templates into its own directory
