+++ templates/index.js
@@ -1,0 +1,7 @@
+const { jsTemplate } = require("./js.template")
+const { tsTemplate } = require("./ts.template")
+
+export {
+	jsTemplate,
+	tsTemplate
+}
